fix: on upgrade, route webhook requests consistently to the new controller deployment
Since we need to fingerprint in the mutating webhook configuration the CA certificate used by the controller, we can get into the situation where webhook requests are routed to controller pods of the previous release version, which are now invalid. We use a digest of the public cert of the controller as label value, and use it in the service to ensure we route only to new pods.