feat: add rootOwnerReference to events
Adopt the k8seventsenricherprocessor to ensure that the events that are
watched by the telemetry-proxy come in with the rootOwnerReference.

Also, filter out events that have involved objects that are not related
with objects that exist because of GVKs we can inject, e.g., filter out
events about custom resources managed by other operators.